Resume: moves file to src/components/resume; index: refactors import path accordingly
