+++ src/components/resume/Resume.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { WebLink } from "./common/WebLink";
+import { WebLink } from "../common/WebLink";
 
 export const Resume = () => {
   return (
