+++ src/index.js
@@ -7,7 +7,7 @@ import { Portfolio } from "./components/portfolio/Portfolio";
 import { EmailForm } from "./components/contact/EmailForm";
 import { Footer } from "./components/layouts/Footer";
 import { NavBar } from "./components/layouts/Navbar";
-import { Resume } from "./components/Resume";
+import { Resume } from "./components/resume/Resume";
 
 import "./css/common.css";
 import "./css/education.css";
